Refactored server command handling
