--- forc-plugins/forc-debug/src/server/handlers/handle_next.rs
@@ -1,8 +1,0 @@
-use crate::server::{AdapterError, DapServer};
-
-impl DapServer {
-    /// Handles a `next` request. Returns true if the server should continue running.
-    pub(crate) fn handle_next(&mut self) -> Result<bool, AdapterError> {
-        self.continue_debugging_tests(true)
-    }
-}
