+++ forc-plugins/forc-debug/src/server/handlers/handle_stack_trace.rs
@@ -1,9 +1,24 @@
 use crate::server::{util, AdapterError, DapServer};
-use dap::types::{StackFrame, StackFramePresentationhint};
+use dap::{
+    responses::ResponseBody,
+    types::{StackFrame, StackFramePresentationhint},
+};
 
 impl DapServer {
     /// Handles a `stack_trace` request. Returns the list of [StackFrame]s for the current execution state.
-    pub(crate) fn handle_stack_trace(&self) -> Result<Vec<StackFrame>, AdapterError> {
+    pub(crate) fn handle_stack_trace_command(
+        &self,
+    ) -> (Result<ResponseBody, AdapterError>, Option<i64>) {
+        let result = self.stack_trace().map(|stack_frames| {
+            ResponseBody::StackTrace(dap::responses::StackTraceResponse {
+                stack_frames,
+                total_frames: None,
+            })
+        });
+        (result, None)
+    }
+
+    fn stack_trace(&self) -> Result<Vec<StackFrame>, AdapterError> {
         let executor = self
             .state
             .executors
